requirements.txt
celery==3.1.18
Django==1.8.4
django-imagekit
python-magic
mysqlclient==1.3.6
PyYAML==3.11
Markdown==2.6.2
